+++ src/components/sideMenu/index.jsx
@@ -9,10 +9,10 @@ export default function SideMenu(){
     return(
         <Style.SideMenu>
             <Style.Container>
-                <Style.Button selected={ window.location.href.indexOf("/time-entry") > -1 } onClick={() => history.push("/time-entry")}>Rota 1</Style.Button>
-                <Style.Button selected={ window.location.href.indexOf("/two") > -1 } onClick={() => history.push("/two")}>Rota 2</Style.Button>
-                <Style.Button selected={ window.location.href.indexOf("/three") > -1 } onClick={() => history.push("/three")}>Rota 3</Style.Button>
-                <Style.Button selected={ window.location.href.indexOf("/four") > -1 } onClick={() => history.push("/four")}>Rota 4</Style.Button>
+                <Style.Button selected={ window.location.href.indexOf("/timeEntry") > -1 } onClick={() => history.push("/timeEntry")}>Apontamento de horas</Style.Button>
+                <Style.Button selected={ window.location.href.indexOf("/dashboards") > -1 } onClick={() => history.push("/dashboards")}>Dashboards</Style.Button>
+                <Style.Button selected={ window.location.href.indexOf("/projectManagement") > -1 } onClick={() => history.push("/projectManagement")}>Gestão de projetos</Style.Button>
+                <Style.Button selected={ window.location.href.indexOf("/register") > -1 } onClick={() => history.push("/register")}>Cadastros</Style.Button>
             </Style.Container> 
         </Style.SideMenu>
     )
