+++ src/components/sideMenu/style.js
@@ -27,13 +27,14 @@ export const Button = styled.div`
     justify-content: flex-start;
     background-color: ${props => props.selected ? 'white' : 'transparent'};
     margin-bottom: 2VH;
-    font-size: 5VH;
+    font-size: 3VH;
     color: ${props => props.selected ? 'black' : 'white'};
     text-align: center;
     line-height: 9.5VH;
     padding: 0 0 0 1VW;
-    border-radius: 50px 0 0 50px;
+    border-radius: 15px 0 0 15px;
     cursor: pointer;
+    transition: all 0.5s;
 
     :hover{
         background-color: ${props => props.selected ? 'white' : '#12075A'}
