+++ src/pages/home/index.css
@@ -23,19 +23,18 @@
 }
 
 .container{
-    display: block;
     position: absolute;
     top: 25VH;
-    left: 18.5VW;
-    width: 43VW;
+    width: 100%;
     height: 60VH;
+    background-color: green;
 }
 
 .subcontainer{
     display: flex;
     justify-content: space-around;
-    width: 43VW;
-    height: 25VH;
+    width: 100%;
+    height: 50%;
     padding-top: 2.5VH;
     padding-bottom: 2.5VH;
 }
@@ -47,11 +46,12 @@
     height: 25VH;
     border-radius: 10px;
     color: white;
-    font-size: 5VH;
-    text-align: center;
-    line-height: 25VH;
+    font-size: 3VH;
+    text-align: left;
+    line-height: 1;
     cursor: pointer;
     transition: all 0.5s;
+    padding: 3% 1%;
 }
 
 .component:hover{
