+++ src/pages/home/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useHistory } from 'react-router-dom'
 import Viewer from '../../Layout/Viewer'
-import './index.css'
+import * as Style from './style'
 
 export default function Home({ setLoad, logged, setLogged }){
 
@@ -16,18 +16,18 @@ export default function Home({ setLoad, logged, setLogged }){
     return(
         <>
             <Viewer>
-                <div class="title">GESPO</div>
-                <div class="subtitle">Gestão de custos e projetos</div>
-                <div class="container">
-                    <div class="subcontainer">
-                        <div class="component" onClick={() => history.push("/time-entry") }>Rota 1</div>
-                        <div class="component" onClick={() => history.push("/two") }>Rota 2</div>
-                    </div>
-                    <div class="subcontainer">
-                        <div class="component" onClick={() => history.push("/three") }>Rota 3</div>
-                        <div class="component" onClick={() => history.push("/four") }>Rota 4</div>
-                    </div>
-                </div>
+                <Style.Title>GESPO</Style.Title>
+                <Style.SubTitle>Gestão de custos e projetos</Style.SubTitle>
+                <Style.Container>
+                    <Style.SubContainer>
+                        <Style.Component onClick={() => history.push("/timeEntry") }>Apontamento de horas</Style.Component>
+                        <Style.Component onClick={() => history.push("/dashboards") }>Dashboards</Style.Component>
+                    </Style.SubContainer>
+                    <Style.SubContainer>
+                        <Style.Component onClick={() => history.push("/projectManagement") }>Gestão de Projetos</Style.Component>
+                        <Style.Component onClick={() => history.push("/register") }>Cadastros</Style.Component>
+                    </Style.SubContainer>
+                </Style.Container>
             </Viewer>
         </>
     );
