+++ src/pages/index.js
@@ -1,10 +1,10 @@
-export { default as Home } from './home'
-export { default as TimeEntry } from './time-entry/component'
-export { default as Two } from './two'
-export { default as Three } from './three'
-export { default as Four } from './four'
 export { default as Login } from './login'
-export { default as HoursProvisioning} from './hoursProvisioning'
-export { default as UserRegister} from './userRegister'
-export { default as CustCenterRegister} from './custCenterRegister'
-export { default as ProjectRegister} from './projectRegister'
+export { default as Home } from './home'
+export { default as TimeEntry } from './time-entry'
+export { default as Dashboards } from './dashboards'
+export { default as ProjectManagement } from './projectManagement'
+export { default as Register } from './register'
+export { default as HoursProvisioning } from './hoursProvisioning'
+export { default as UserRegister } from './userRegister'
+export { default as CustCenterRegister } from './custCenterRegister'
+export { default as ProjectRegister } from './projectRegister'
