+++ src/pages/register/style.js
