+++ src/routes.js
@@ -18,14 +18,14 @@ export default function Routes({ setLoad }){
                     <Pages.Login setLoad={setLoad}
                                  logged={logged} 
                                  setLogged={setLogged} />} exact />
-                <Route path="/time-entry" render={(props) => <Pages.TimeEntry setLoad={setLoad} />} exact />
-                <Route path="/two" render={(props) => <Pages.Two setLoad={setLoad} />} exact />
-                <Route path="/three" render={(props) => <Pages.Three setLoad={setLoad} />} exact />
-                <Route path="/four" render={(props) => <Pages.Four setLoad={setLoad} />} exact />
-                <Route path="/hoursProvisioning" render={(props) => <Pages.HoursProvisioning setLoad={setLoad} />} exact />
-                <Route path="/UserRegister" render={(props) => <Pages.UserRegister setLoad={setLoad} />} exact />
-                <Route path="/ProjectRegister" render={(props) => <Pages.ProjectRegister setLoad={setLoad} />} exact />
-                <Route path="/CustCenterRegister" render={(props) => <Pages.CustCenterRegister setLoad={setLoad} />} exact />
+                <Route path="/timeEntry" render={(props) => <Pages.TimeEntry setLoad={setLoad} />} exact />
+                <Route path="/dashboards" render={(props) => <Pages.Dashboards setLoad={setLoad} />} exact />
+                <Route path="/projectManagement" render={(props) => <Pages.ProjectManagement setLoad={setLoad} />} exact />
+                <Route path="/projectManagement/hoursProvisioning" render={(props) => <Pages.HoursProvisioning setLoad={setLoad} />} exact />
+                <Route path="/register" render={(props) => <Pages.Register setLoad={setLoad} />} exact />
+                <Route path="/register/UserRegister" render={(props) => <Pages.UserRegister setLoad={setLoad} />} exact />
+                <Route path="/register/ProjectRegister" render={(props) => <Pages.ProjectRegister setLoad={setLoad} />} exact />
+                <Route path="/register/CustCenterRegister" render={(props) => <Pages.CustCenterRegister setLoad={setLoad} />} exact />
             </Switch>
         </BrowserRouter>
     );
